Some fixes in EMG measurements
Plotting can now be switched off (for example when there is no Python installed on the system).
Plotting has been simplified in PythonPlotter.
Readability has been improved slightly in DelsysDataReader.
CMakeLists has been updated to allow preprocessor definitions that disable Python-related classes and methods if Python is not found on the system.
The main program of Delsys testing has been updated to handle quaternion and EMG retrieval with multithreading.