Updated EMG handling
Changed DelsysDataReader so that it no longer has anything to do with PythonPlotter directly. Previously PythonPlotter methods were invoked from DelsysDataReader, now DelsysDataReader is only used to fetch the EMG data and it is fed to PythonPlotter separately in the main program.
Updated test_EMG and test_Delsys to reflect the way DelsysDataReader and PythonPlotter currently work.
Fixed a typo in CMakeLists.txt.