Improved EMG reading
Client.h and Client.cpp have been modified to that RecvBytes() is waits for 64000 bytes only if an optional argument hasn't been specified. Previously RecvBytes() always waited for 64000 bytes, which resulted in slow performance.
In DelsysDataReader, preassigned char array size to be 4 for byte-to-float conversion. Modified updateEMGData() to read smaller amount of bytes and simplified the code.