Made PointTracker output communication more flexible and added a config option to forgo rotation conversion to KUKA
The container for mirrored point location and rotation is no longer a fixed-size array, but a vector, and is only converted to an array when it is sent to the Java client. This allows us to send only the amount of doubles that are required (6 if rotations are Euler angles, 7 if quaternions).
Additionally, added parameter "transform_rotations_to_kuka" in MainConfiguration.xml so the user can toggle whether we perform conversion from OpenSim coordinate system to KUKA iiwa coordinate system before sending the data to the Java client.